Allow deriving Into for primitive enums
This derive is only allowed when all of the following are true:
1. The derive is on an enum.
2. The enum has an explicit repr (in an aim to show the deriver has
   considered what type this enum should be convertable into).
3. The repr is a fixed-size integer type (i.e. not `C`).
4. None of the enum's variants may hold data.
5. None of the enum's variants are declared as holding an empty tuple as
   data (because for `enum E { Empty() }`, `E::Empty as isize` converts
   the address of the function pointer, rather than the discriminant).
6. None of the enum's variants are declared as holding empty named
   values (i.e. `enum E { Empty{} }`) for consistency.

Currently this tends to be done using `as`, which has two flaws:
1. It will silently truncate. e.g. if you have a `repr(u64)` enum, and
   you write `value as u8` (perhaps because the repr changed), you get
   no warning or error that truncation may occur.
2. You cannot use enums in code which is generic over `From` or `Into`.

Other types, such as primitive numeric types, support both `as` casts
and `From`/`Into` conversions, so making this trivial for primitive
enums too seems reasonable.

There are a number of crates which provide macros to provide this
implementation, as well as `TryFrom` implementations. The advantages of
using infallible conversions, rather than possibly-truncating ones feels
like it warrants a place in `std`.