Allow conversion of primitive enums to their repr
Currently this tends to be done using `as`, which has some flaws:
1. It will silently truncate. e.g. if you have a `repr(u64)` enum, and
   you write `value as u8` (perhaps because the repr changed), you get
   no warning or error that truncation will occur.
2. You cannot use enums in code which is generic over `From` or `Into`.
3. Not all enums _want_ to expose conversion to their underlying repr as
   part of their public API.

Instead, by allowing an `IntoUnderlying` implementation to be easily
derived, we can avoid these issues.

There are a number of crates which provide macros to provide this
implementation, as well as `TryFrom` implementations. I believe this is
enough of a foot-gun that it should be rolled into `std`.

See https://github.com/rust-lang/rfcs/pull/3046 for more information.